+++ frontend/src/components/auth/Logout.svelte
