+++ frontend/src/components/news/NewsCard.svelte
