+++ frontend/src/components/news/NewsPanel.svelte
